+++ backend/src/main/java/com/onair/hearit/auth/presentation/AuthController.java
@@ -1,13 +1,15 @@
 package com.onair.hearit.auth.presentation;
 
 import com.onair.hearit.auth.application.AuthService;
-import com.onair.hearit.auth.dto.UserContext;
-import com.onair.hearit.auth.dto.request.KakaoLoginRequest;
+import com.onair.hearit.auth.application.OAuthProvider;
+import com.onair.hearit.auth.domain.UserContext;
 import com.onair.hearit.auth.dto.request.LoginRequest;
+import com.onair.hearit.auth.dto.request.OAuthLoginRequest;
 import com.onair.hearit.auth.dto.request.SignupRequest;
 import com.onair.hearit.auth.dto.request.TokenReissueRequest;
 import com.onair.hearit.auth.dto.response.LoginTokenResponse;
 import com.onair.hearit.auth.dto.response.TokenReissueResponse;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -33,8 +35,8 @@ public class AuthController {
     }
 
     @PostMapping("/kakao-login")
-    public ResponseEntity<LoginTokenResponse> loginOrSignupWithKakao(@RequestBody KakaoLoginRequest request) {
-        LoginTokenResponse response = authService.loginOrSignupWithKakao(request);
+    public ResponseEntity<LoginTokenResponse> loginOrSignupWithKakao(@Valid @RequestBody OAuthLoginRequest request) {
+        LoginTokenResponse response = authService.loginOrSignUp(request, OAuthProvider.KAKAO);
         return ResponseEntity.ok(response);
     }
