+++ backend/src/test/java/com/onair/hearit/auth/infrastructure/oauth/kakao/KakaoOAuthServiceTest.java
@@ -1,6 +1,6 @@
-package com.onair.hearit.auth.infrastructure.client;
+package com.onair.hearit.auth.infrastructure.oauth.kakao;
 
-import com.onair.hearit.auth.dto.response.KakaoUserInfoResponse;
+import com.onair.hearit.auth.dto.request.OAuthUserInfo;
 import com.onair.hearit.fixture.IntegrationTest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Disabled;
@@ -8,18 +8,17 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-class KakaoUserInfoClientTest extends IntegrationTest {
+class KakaoOAuthServiceTest extends IntegrationTest {
 
     @Autowired
-    KakaoUserInfoClient kakaoUserInfoClient;
+    KakaoOAuthService kakaoOAuthService;
 
     @Disabled("실제 카카오 외부 API를 호출하는 테스트-필요시 활성화")
     @Test
     @DisplayName("실제 카카오 API를 호출하여 사용자정보를 가져온다.")
     void fetchUserInfo_usingRealKakaoApi() {
         String kakaoAccessToken = "yC5CmZ1kj1jHcCICRBWX_dvbXIMqfn8rAAAAAQoXBi4AAAGYE03XhpQkbXeV0h_w";
-
-        KakaoUserInfoResponse response = kakaoUserInfoClient.getUserInfo(kakaoAccessToken);
+        OAuthUserInfo response = kakaoOAuthService.fetchUser(kakaoAccessToken);
 
         SoftAssertions.assertSoftly(softly -> {
             softly.assertThat(response).isNotNull();
